% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnshelpers-package.R
\docType{package}
\name{dnshelpers}
\alias{dnshelpers}
\alias{dnshelpers-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/dnshelpers}
\item BugReports: \url{https://gitlab.com/hrbrmstr/dnshelpers/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
