% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-dmarc.R
\name{parse_dmarc}
\alias{parse_dmarc}
\title{Parse DMARCv1 DNS data reponses into a data frame}
\usage{
parse_dmarc(x)
}
\arguments{
\item{x}{a character vector of DMARCv1 responses}
}
\value{
data frame (tibble)
}
\description{
Parse DMARCv1 DNS data reponses into a data frame
}
