% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-mta-sts.R
\name{parse_mta_sts}
\alias{parse_mta_sts}
\title{Parse MTA-STS DNS data reponses into a list}
\usage{
parse_mta_sts(x)
}
\arguments{
\item{x}{a character vector of MTA-STS responses}
}
\value{
list (that will collape nicely into a data frame with \code{dply::bind_rows()}
or \code{data.table::rbindlist(..., fill=TRUE)}).
}
\description{
Parse MTA-STS DNS data reponses into a list
}
