% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-mx.R
\name{parse_mx}
\alias{parse_mx}
\title{Parse MX DNS data reponses into a data frame}
\usage{
parse_mx(x)
}
\arguments{
\item{x}{a character vector of MX responses}
}
\value{
data frame (tibble)
}
\description{
Parse MX DNS data reponses into a data frame
}
