% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-soa.R
\name{parse_soa}
\alias{parse_soa}
\title{Parse SOA DNS data reponses into a data frame}
\usage{
parse_soa(x)
}
\arguments{
\item{x}{a character vector of SOA responses}
}
\value{
data frame (tibble)
}
\description{
Parse SOA DNS data reponses into a data frame
}
