% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-spf.R
\name{parse_spf}
\alias{parse_spf}
\title{Parse SPF TXT DNS data reponses into a list}
\usage{
parse_spf(x)
}
\arguments{
\item{x}{a character vector of DMARCv1 responses}
}
\value{
list (that will collape nicely into a data frame with \code{dply::bind_rows()}
or \code{data.table::rbindlist(..., fill=TRUE)}).
}
\description{
Parse SPF TXT DNS data reponses into a list
}
